+++ Engine/Goal.h
@@ -6,22 +6,22 @@ class Goal
 {
 public:
 	void Bar( Color c, Graphics& gfx);
-	void Respawn(int x_in, int y_in);
+	void Respawn(float x_in, float y_in);
 	void ProcessConsumption(const Dude& dude);
 	void ChangeColor();
 	void Draw(Graphics& gfx) const
 	{
-		gfx.DrawRectDim(x, y, dimension, dimension, c);
+		gfx.DrawRectDim(int(x), int(y), int(dimension), int(dimension), c);
 	}
 	bool GetTeleport() const;
 private:
-	Color c = { 255, 255, 150};
-	int x;
-	int y;
+	Color c = { 255, 255, 150 };
+	float x;
+	float y;
 	int progress = 0;
 	bool teleport = true;
 	bool Cchange = true;
 	bool StartAdd = false;
-	static constexpr int dimension = 20;
+	static constexpr float dimension = 20.0f;
 };
